Refactor strategies into reuseable component
