+++ src/helpers/interfaces.ts
@@ -58,11 +58,7 @@ export interface ExtendedSpace {
   name: string;
   symbol: string;
   network: string;
-  strategies: {
-    name: string;
-    network: string;
-    params: Record<string, unknown>;
-  }[];
+  strategies: SpaceStrategy[];
   about: string;
   avatar: string;
   skin: string;
@@ -86,6 +82,12 @@ export interface ExtendedSpace {
     quorum: number | null;
     type: string | null;
   };
+}
+
+export interface SpaceStrategy {
+  name: string;
+  network: string;
+  params: Record<string, unknown>;
 }
 
 export interface Proposal {
@@ -112,7 +114,7 @@ export interface Proposal {
   votes: number;
   plugins: Record<string, any>;
   space: { id: string; name: string };
-  strategies: { name: string; network: string; params: Record<string, any> }[];
+  strategies: SpaceStrategy[];
 }
 
 export interface Results {
